fix(theme): don't let navbar obstruct clicks to top part of scrollbar
On desktop screen sizes, if the sidebar has a scrollbar, the very top
part of the scrollbar can't be clicked on because the navbar obstructs
it due to having a higher z-index. (This higher z-index is necessary so
that the navbar is visible.) With this fix, the navbar container has
`pointer-events: none;` so that clicks to the scrollbar underneath will
work. Child elements of the navbar that need to receive clicks have
`pointer-events: all;` turned on for them.

fix vueuse/vueuse#2086