fix: fix build for vite 3 + "type": "module"
Background: pnpm injects `NODE_PATH` when installing npm script binaries
in order to simulate flat install structure when running npm scripts.
This previously made files outside of VitePress to be able to import
transitive deps (e.g. `vue`), but this breaks when upgrading to Vite 3
or in esm mode, because:

- "type": "module", aka ESM mode doesn't support `NODE_PATH`, so now
  project files can't resolve `vue` which is a transitive dep.

- Vite 3 now auto-resolves SSR externals, but it requires the
  dep to be resolvable first. Since it can't resovle `vue`, the Rollup
  build will fail.

The fix: detect if `vue` is resolvable from project root's node_modules.
If not, create a symlink to the version of `vue` from VitePress' own
deps.