perf: inline site data on page
Previously the site data is bundled in JavaScript, and when it changes,
it invalidates the chunk hash of the framework chunk, and in turn
invlidates the chunk hash of **every page**.

With this change, the site data is now inlined in each page as a JSON
string, similar to the page chunk hash data. This ensures that config
changes will no longer affect JavaScript chunk hashes.