feat(theme): containers: add the success custom block
This adds a custom block with the `success` keyword, which is a green
container.

The colours for this container already exist in vars.css
(--vp-c-success*), which, before this commit, is completely unused
outside of generating a diff.

For example, this can be used for messages such as:
```
::: success

You have completed the walkthrough!

:::
```